+++ src/app/product/product-list/product-list.component.ts
@@ -20,7 +20,7 @@ export class ProductListComponent implements OnInit {
   ) { }
 
   onBuyProduct(product: Product): void {
-    this.cartService.addToCart(product);
+    this.cartService.addProduct(product);
   }
 
   ngOnInit() {