+++ src/app/shared/font-click.directive.spec.ts
@@ -1,0 +1,8 @@
+import { FontClickDirective } from './font-click.directive';
+
+describe('FontClickDirective', () => {
+  it('should create an instance', () => {
+    const directive = new FontClickDirective();
+    expect(directive).toBeTruthy();
+  });
+});