+++ changelog.md
@@ -20,3 +20,13 @@ All notable changes to this project will be documented in this file.
 - Product structure
 - Cart list structure
 - Product model from interface to class
+
+## 2020-02-05
+### Added
+- Fields with content, quantity and price to CartService
+
+### Changed
+- Removed redundant FirstComponent
+- Product structure
+- Cart list structure
+- Product model from interface to class